# Conflicts:
#	src/main/java/com/wishlist/cst438project2/controller/ItemController.java
#	src/main/java/com/wishlist/cst438project2/document/User.java
#	src/main/java/com/wishlist/cst438project2/dto/UserDTO.java

+++ src/main/java/com/wishlist/cst438project2/controller/ItemController.java
@@ -2,11 +2,13 @@ package com.wishlist.cst438project2.controller;
 
 import com.wishlist.cst438project2.document.Item;
 import com.wishlist.cst438project2.dto.CreateItemDTO;
+import com.wishlist.cst438project2.common.Constants;
+import com.wishlist.cst438project2.common.TokenManager;
 import com.wishlist.cst438project2.dto.ItemDTO;
+import com.wishlist.cst438project2.dto.UserTokenDTO;
 import com.wishlist.cst438project2.exception.BadRequestException;
-import com.wishlist.cst438project2.integration.FirebaseIntegration;
+import com.wishlist.cst438project2.exception.UnauthorizedException;
 import com.wishlist.cst438project2.service.ItemService;
-
 import java.util.List;
 import java.util.Objects;
 import lombok.extern.slf4j.Slf4j;
@@ -32,16 +34,21 @@ public class ItemController {
     @Autowired
     private ItemService itemService;
 
+    @Autowired
+    private TokenManager tokenManager;
+
     /**
      * POST request to create an item in the database.
      * returns item creation timestamp
-     * referenced:
-     *     - pulling mutiple variables in request body: https://stackoverflow.com/questions/53332930/error-when-performing-post-operation-in-spring
      */
     @PostMapping
-    public String createItem(@RequestBody CreateItemDTO createItemDTO) {
+    public String createItem(@RequestHeader String accessToken, @RequestBody CreateItemDTO createItemDTO) {
         log.info("ItemController: Starting createItem");
         try {
+            UserTokenDTO userTokenDTO = tokenManager.getUser(accessToken);
+            if(Objects.isNull(userTokenDTO))
+                throw new UnauthorizedException(Constants.ERROR_INVALID_TOKEN);
+
             if (Objects.isNull(createItemDTO.getItemDTO()) || (createItemDTO.getItemDTO().getName().isBlank())) {
                 throw new BadRequestException();
             } else {
@@ -64,9 +71,14 @@ public class ItemController {
      * returns List<ItemDTO> collection
      */
     @GetMapping
-    public List<ItemDTO> getAllItems() {
+    public List<ItemDTO> getAllItems(@RequestHeader String accessToken) {
         log.info("ItemController: Starting getAllItems");
         try {
+
+            UserTokenDTO userTokenDTO = tokenManager.getUser(accessToken);
+            if(Objects.isNull(userTokenDTO))
+                throw new UnauthorizedException(Constants.ERROR_INVALID_TOKEN);
+
             List<ItemDTO> collection = itemService.getAllItems();
             log.info("ItemController: Exiting successful getAllItems");
             return collection;
@@ -80,12 +92,16 @@ public class ItemController {
      * DELETE request to remove the item associated with a given user ID and item name
      * returns timestamp of successful deletion
      */
-    @DeleteMapping
-    public String removeItem(@RequestParam String item_name, @RequestParam int userId) {
+    @RequestMapping(method = RequestMethod.DELETE, params = {"item_name", "userId"}, headers = "accessToken")
+    public String removeItem(@RequestHeader String accessToken, @RequestParam String item_name, @RequestParam int userId) {
         // TODO: adjust from removeItem onward to take String username --> ItemServiceImpl: convert username to userId via firebaseIntegration.getUserId(username)
         log.info("ItemController: Starting removeItem");
         log.info(String.format("ItemController: removeItem:\n    name: %s\n    userId: %s", item_name, userId));
-        // TODO: add item delete confirmation message --> I think that's front end
+
+        UserTokenDTO userTokenDTO = tokenManager.getUser(accessToken);
+        if(Objects.isNull(userTokenDTO))
+            throw new UnauthorizedException(Constants.ERROR_INVALID_TOKEN);
+
         String timestamp = itemService.removeItem(item_name, userId);
         return timestamp;
     }
@@ -95,9 +111,14 @@ public class ItemController {
      * returns timestamp of successful update
      */
     @PatchMapping
-    public String updateItem(@RequestParam String item_name, @RequestBody ItemDTO updatedItemDTO) {
+    public String updateItem(@RequestHeader String accessToken, @RequestParam String item_name, @RequestBody ItemDTO updatedItemDTO) {
         log.info("ItemController: Starting updateItem");
         log.info(String.format("ItemController: updateItem:\n    name: %s\n    userId: %s", item_name, updatedItemDTO.getUserId()));
+
+        UserTokenDTO userTokenDTO = tokenManager.getUser(accessToken);
+        if(Objects.isNull(userTokenDTO))
+            throw new UnauthorizedException(Constants.ERROR_INVALID_TOKEN);
+
         String timestamp = itemService.updateItem(item_name, updatedItemDTO);
         return timestamp;
     }
@@ -106,11 +127,16 @@ public class ItemController {
      * GET request to retrieve a list of items given a user's id as list identifier
      * returns list of items
      */
-    @RequestMapping(method = RequestMethod.GET, params = "list")
+    @RequestMapping(method = RequestMethod.GET, params = "list", headers = "accessToken")
     // https://stackoverflow.com/a/43546809
-    public List<ItemDTO> getUserItems(@RequestParam("list") int userId) {
+    public List<ItemDTO> getUserItems(@RequestHeader String accessToken, @RequestParam("list") int userId) {
         log.info("ItemController: Starting getUserItems");
         log.info(String.format("ItemController: getUserItems:\n     userId: %s", userId));
+
+        UserTokenDTO userTokenDTO = tokenManager.getUser(accessToken);
+        if(Objects.isNull(userTokenDTO))
+            throw new UnauthorizedException(Constants.ERROR_INVALID_TOKEN);
+
         List<ItemDTO> userItems = itemService.getUserItems(userId);
         return userItems;
     }
@@ -119,12 +145,37 @@ public class ItemController {
      * GET request to retrieve list of items matching given keywords
      * returns list of keyword relevant items
      */
-    @RequestMapping(method = RequestMethod.GET, params = "search")
-    public List<ItemDTO> getSearchAllItems(@RequestParam("search") List<String> keywords) {
+    @RequestMapping(method = RequestMethod.GET, params = "search", headers = "accessToken")
+    public List<ItemDTO> getSearchAllItems(@RequestHeader String accessToken, @RequestParam("search") List<String> keywords) {
         log.info("ItemController: Starting getSearchItems");
         log.info(String.format("ItemController: getSearchItems:\n     keywords: %s", keywords));
+
+        UserTokenDTO userTokenDTO = tokenManager.getUser(accessToken);
+        if(Objects.isNull(userTokenDTO))
+            throw new UnauthorizedException(Constants.ERROR_INVALID_TOKEN);
+
         // TODO: add functionality to ItemService, ItemServiceImpl, and FirebaseIntegration
         List<ItemDTO> userItems = itemService.getSearchAllItems(keywords);
         return userItems;
+    }
+
+    /**
+     * DELETE request to remove all items associated with given userId upon account deletion
+     * returns timestamp of successful deletion
+     */
+    // TODO: remove mapping! I think this is ONLY supposed to be a helper for the delete user account request mapping.
+    // TODO: removeItemsByUser mapping is for testing purposed only!
+    @RequestMapping(method = RequestMethod.DELETE, params = "userId", headers = "accessToken")
+    public String removeItemsByUser(@RequestHeader String accessToken, @RequestParam int userId) {
+        log.info("ItemController: Starting removeItemsByUser");
+        log.info(String.format("ItemController: removeItemsByUser:\n    userId: %s", userId));
+
+        UserTokenDTO userTokenDTO = tokenManager.getUser(accessToken);
+        if(Objects.isNull(userTokenDTO))
+            throw new UnauthorizedException(Constants.ERROR_INVALID_TOKEN);
+
+        String timestamp = itemService.removeItemsByUser(userId);
+        log.info("ItemController: Exiting removeItemsByUser");
+        return timestamp;
     }
 }
