+++ src/main/java/com/wishlist/cst438project2/controller/UserController.java
@@ -1,15 +1,17 @@
 package com.wishlist.cst438project2.controller;
 
 import com.wishlist.cst438project2.common.Constants;
+import com.wishlist.cst438project2.common.TokenManager;
 import com.wishlist.cst438project2.common.Utils;
 import com.wishlist.cst438project2.dto.*;
 import com.wishlist.cst438project2.exception.BadRequestException;
+import com.wishlist.cst438project2.exception.UnauthorizedException;
 import com.wishlist.cst438project2.service.UserService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 
-import javax.validation.Valid;
 import java.util.Objects;
 
 @RestController
@@ -19,6 +21,9 @@ public class UserController {
 
     @Autowired
     private UserService userService;
+
+    @Autowired
+    private TokenManager tokenManager;
 
     /**
      * This API is used for user registration
@@ -53,11 +58,16 @@ public class UserController {
      * @return user creation timestamp
      */
     @PutMapping("/updateUser")
-    public UserDTO updateUser(@RequestBody UserDTO userDTO) {
+    public UserDTO updateUser(@RequestHeader String accessToken, @RequestBody UserDTO userDTO) {
 
         log.info("UserController: Starting updateUser");
 
         try {
+
+            UserTokenDTO userTokenDTO = tokenManager.getUser(accessToken);
+
+            if(Objects.isNull(userTokenDTO))
+                throw new UnauthorizedException(Constants.ERROR_INVALID_TOKEN);
 
             if(Objects.isNull(userDTO))
                 throw new BadRequestException();
@@ -80,61 +90,97 @@ public class UserController {
      * @return message if the password was changed successfully or returns error message
      */
     @PutMapping("/changePassword")
-    public String changePassword(@RequestBody ChangePasswordDTO changePasswordDTO) {
+    public ResponseDTO<Long> changePassword(@RequestHeader String accessToken, @RequestBody ChangePasswordDTO changePasswordDTO) {
 
         log.info("UserController: Starting changePassword");
 
+        ResponseDTO<Long> responseDTO = new ResponseDTO<>();
+        String msg = "";
+        UserTokenDTO userTokenDTO;
+        int httpStatusCode = HttpStatus.INTERNAL_SERVER_ERROR.value();
         try {
+
+            userTokenDTO = tokenManager.getUser(accessToken);
+
+            if(Objects.isNull(userTokenDTO))
+                throw new UnauthorizedException(Constants.ERROR_INVALID_TOKEN);
 
             if(Objects.isNull(changePasswordDTO))
                 throw new BadRequestException();
 
-            boolean isValid = Utils.validatePassword(changePasswordDTO.getPassword(), changePasswordDTO.getConfirmPassword());
+            boolean isValid = Utils.validatePassword(changePasswordDTO.getNewPassword(), changePasswordDTO.getConfirmPassword());
 
-            String msg = "";
             if(isValid) {
-                msg = userService.changePassword(changePasswordDTO);
+                msg = userService.changePassword(userTokenDTO.getUsername(), changePasswordDTO);
+                httpStatusCode = HttpStatus.OK.value();
             } else {
                 msg = Constants.ERROR_USER_PASSWORD_MISMATCH;
             }
-            log.info("UserController: Exiting changePassword");
 
-            return msg;
+            responseDTO.setStatus(httpStatusCode);
 
         } catch (Exception ex) {
             log.error(ex.getMessage(), ex);
+
+            if(ex instanceof UnauthorizedException)
+                httpStatusCode = HttpStatus.UNAUTHORIZED.value();
+            else if (ex instanceof BadRequestException)
+                httpStatusCode = HttpStatus.BAD_REQUEST.value();
+
+            responseDTO.setStatus(httpStatusCode);
             throw ex;
         }
+
+        log.info("UserController: Exiting changePassword");
+
+        responseDTO.setData(userTokenDTO.getUserId());
+        responseDTO.setMessage(msg);
+        return responseDTO;
     }
 
     @PostMapping("/login")
-    public String login(@RequestBody SignInDTO signInDTO) {
+    public ResponseDTO<UserLoginDTO> login(@RequestBody SignInDTO signInDTO) {
 
         log.info("UserController: Starting login");
 
+        ResponseDTO<UserLoginDTO> responseDTO = new ResponseDTO<>();
+        UserLoginDTO userLoginDTO = null;
+        int httpStatusCode;
         try {
 
             if (Objects.isNull(signInDTO))
                 throw new BadRequestException();
 
-            String responseTimestamp = userService.login(signInDTO);
+            userLoginDTO = userService.login(signInDTO);
+            if(Objects.isNull(userLoginDTO))
+                throw new UnauthorizedException();
 
+            httpStatusCode = HttpStatus.OK.value();
             log.info("UserController: Exiting login");
-
-            return responseTimestamp;
 
         } catch (Exception ex) {
             log.error(ex.getMessage(), ex);
-            throw ex;
+            httpStatusCode = HttpStatus.UNAUTHORIZED.value();
         }
+
+        responseDTO.setStatus(httpStatusCode);
+        responseDTO.setData(userLoginDTO);
+        responseDTO.setMessage(Constants.USER_LOGIN_SUCCESSFUL);
+
+        return responseDTO;
     }
 
     @DeleteMapping("/deleteUser")
-    public String deleteUser(@RequestBody DeleteUserDTO deleteUserDTO) {
+    public String deleteUser(@RequestHeader String accessToken, @RequestBody DeleteUserDTO deleteUserDTO) {
 
         log.info("UserController: Starting deleteUser");
 
         try{
+
+            UserTokenDTO userTokenDTO = tokenManager.getUser(accessToken);
+
+            if(Objects.isNull(userTokenDTO))
+                throw new UnauthorizedException(Constants.ERROR_INVALID_TOKEN);
 
             userService.deleteUser(deleteUserDTO);
 
@@ -145,6 +191,29 @@ public class UserController {
         } catch (Exception ex) {
             log.error(ex.getMessage(), ex);
             throw ex;
+        }
+    }
+
+    @PostMapping("/logout")
+    public String logout(@RequestHeader String accessToken) {
+
+        log.info("UserController: Starting logout");
+
+        try {
+            UserTokenDTO userTokenDTO = tokenManager.getUser(accessToken);
+
+            if(Objects.isNull(userTokenDTO))
+                throw new UnauthorizedException(Constants.ERROR_INVALID_TOKEN);
+
+            userService.logout(accessToken);
+
+            log.info("UserController: Exiting logout");
+
+            return Constants.USER_LOGOUT_SUCCESSFUL;
+
+        } catch (Exception ex) {
+            log.error(ex.getMessage(), ex);
+            throw new UnauthorizedException();
         }
     }
 }
